+++ website/lib/call_to_action.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 const _breakpoint = 768;
 
@@ -26,7 +27,7 @@ class CallToAction extends StatelessWidget {
             ),
           ),
           const SizedBox(height: 29),
-          const _DownloadButton(),
+          const _DocumentationButton(),
           SizedBox(height: isNarrowScreen ? 60 : 104),
         ],
       ),
@@ -34,8 +35,8 @@ class CallToAction extends StatelessWidget {
   }
 }
 
-class _DownloadButton extends StatelessWidget {
-  const _DownloadButton();
+class _DocumentationButton extends StatelessWidget {
+  const _DocumentationButton();
 
   @override
   Widget build(BuildContext context) {
@@ -43,7 +44,9 @@ class _DownloadButton extends StatelessWidget {
 
     return MaterialButton(
       color: const Color(0xFFFAE74F),
-      onPressed: () {},
+      onPressed: () => launch(
+        'https://github.com/superlistapp/super_editor/blob/main/super_editor/README.md',
+      ),
       padding: isNarrowScreen
           ? const EdgeInsets.symmetric(horizontal: 32, vertical: 20)
           : const EdgeInsets.symmetric(horizontal: 48, vertical: 24),
@@ -51,7 +54,7 @@ class _DownloadButton extends StatelessWidget {
       height: 42,
       elevation: 0,
       child: const Text(
-        'Download Now',
+        'Documentation',
         style: TextStyle(
           fontWeight: FontWeight.bold,
           fontSize: 29,
